+++ Extension/src/pages/options/components/Allowlist/Allowlist.jsx
@@ -40,9 +40,10 @@ import { translator } from '../../../../common/translators/translator';
 import { OptionsPageSections } from '../../../../common/nav';
 import { exportData, ExportTypes } from '../../../common/utils/export';
 import { getFirstNonDisabledElement } from '../../../common/utils/dom';
-import { DynamicRulesLimitsWarning } from '../Warnings';
+import { DynamicRulesLimitsWarning, ClipboardPermissionWarning } from '../Warnings';
 import { SavingFSMState, CURSOR_POSITION_AFTER_INSERT } from '../../../common/components/Editor/savingFSM';
 import { usePreventUnload } from '../../../common/hooks/usePreventUnload';
+import { UserAgent } from '../../../../common/user-agent';
 
 import { AllowlistSavingButton } from './AllowlistSavingButton';
 import { AllowlistSwitcher } from './AllowlistSwitcher';
@@ -218,6 +219,7 @@ const Allowlist = observer(() => {
                 inlineControl={(<AllowlistSwitcher labelId={switchTitleId} />)}
             />
             <DynamicRulesLimitsWarning />
+            {UserAgent.isFirefoxMobile && <ClipboardPermissionWarning />}
             <Editor
                 name="allowlist"
                 editorRef={editorRef}