+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Made clipboard permissions optional for Firefox, with a prompt shown when needed in User Rules and Allowlist editors [#3364]
+
 ### Changed
 
 - Updated [@adguard/agtree] to v3.3.1.
@@ -14,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated [@adguard/scriptlets] to v2.2.13.
 - Updated [@adguard/tsurlfilter] to v3.5.1.
 - Updated [@adguard/tswebextension] to v3.2.16.
+
+[#3364]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3364
 
 ## [5.2.400] - 2025-11-24